backend/requirements.txt
aiofiles==23.1.0
aniso8601==7.0.0
annotated-types==0.7.0
anyio==4.7.0
async-exit-stack==1.0.1
async-generator==1.10
certifi==2024.12.14
click==8.1.7
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.5
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
numpy
orjson==3.10.15
pandas==2.2.3
psycopg2==2.9.10
pydantic==2.10.3
pydantic-core==2.27.1
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pyyaml==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.36
starlette==0.41.3
typer==0.15.1
typing-extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
